feat: workspace dropdown

+++ ui/src/views/workspace/index.vue
@@ -6,29 +6,25 @@
         <div class="workspace-left border-r p-16">
           <div class="workspace-left_title">
             <h4 class="medium">{{ $t('views.workspace.list') }}</h4>
-            <el-tooltip effect="dark"
-                        :content="`${$t('common.create')}${$t('views.workspace.title')}`"
-                        placement="top">
+            <el-tooltip effect="dark" :content="`${$t('common.create')}${$t('views.workspace.title')}`" placement="top">
               <el-button type="primary" text @click="createOrUpdateWorkspace()">
                 <AppIcon iconName="app-copy"></AppIcon>
               </el-button>
             </el-tooltip>
           </div>
           <div class="p-8">
-            <el-input v-model="filterText" :placeholder="$t('common.search')" prefix-icon="Search"
-                      clearable/>
+            <el-input v-model="filterText" :placeholder="$t('common.search')" prefix-icon="Search" clearable />
           </div>
           <div class="list-height-left">
             <el-scrollbar v-loading="loading">
-              <common-list :data="filterList" @click="clickWorkspace"
-                           :default-active="currentWorkspace?.id">
+              <common-list :data="filterList" @click="clickWorkspace" :default-active="currentWorkspace?.id">
                 <template #default="{ row }">
                   <div class="flex-between">
                     <span>{{ row.name }}</span>
                     <el-dropdown :teleported="false">
                       <el-button text>
                         <el-icon class="color-secondary">
-                          <MoreFilled/>
+                          <MoreFilled />
                         </el-icon>
                       </el-button>
                       <template #dropdown>
@@ -62,30 +58,29 @@
         <div class="workspace-right" v-loading="loading">
           <div class="flex align-center" style="margin-bottom: 20px;">
             <h4 class="medium">{{ currentWorkspace?.name }}</h4>
-            <el-divider direction="vertical" class="mr-8 ml-8"/>
-            <AppIcon iconName="app-wordspace" style="font-size: 16px"
-                     class="color-input-placeholder"></AppIcon>
+            <el-divider direction="vertical" class="mr-8 ml-8" />
+            <AppIcon iconName="app-wordspace" style="font-size: 16px" class="color-input-placeholder"></AppIcon>
             <span class="color-input-placeholder ml-4">
               {{ currentWorkspace?.user_count }}
             </span>
           </div>
-          <Member :currentWorkspace="currentWorkspace"/>
+          <Member :currentWorkspace="currentWorkspace" />
         </div>
       </div>
     </el-card>
 
-    <CreateOrUpdateWorkspaceDialog ref="createOrUpdateWorkspaceDialogRef" @refresh="refresh"/>
+    <CreateOrUpdateWorkspaceDialog ref="createOrUpdateWorkspaceDialogRef" @refresh="refresh" />
   </div>
 </template>
 
 <script lang="ts" setup>
-import {onMounted, ref, watch} from 'vue'
+import { onMounted, ref, watch } from 'vue'
 import WorkspaceApi from '@/api/workspace'
-import {t} from '@/locales'
+import { t } from '@/locales'
 import Member from './component/Member.vue'
 import CreateOrUpdateWorkspaceDialog from './component/CreateOrUpdateWorkspaceDialog.vue'
-import type {WorkspaceItem} from '@/api/type/workspace'
-import {MsgSuccess, MsgConfirm} from '@/utils/message'
+import type { WorkspaceItem } from '@/api/type/workspace'
+import { MsgSuccess, MsgConfirm } from '@/utils/message'
 
 const filterText = ref('')
 const loading = ref(false)
@@ -137,25 +132,43 @@ function createOrUpdateWorkspace(item?: WorkspaceItem) {
   createOrUpdateWorkspaceDialogRef.value?.open(item);
 }
 
-// TODO 该工作空间下存在 知识库资源、应用资源，无法删除
-function deleteWorkspace(item: WorkspaceItem) {
-  MsgConfirm(
-    `${t('views.workspace.delete.confirmTitle')}${item.name} ?`,
-    t('views.workspace.delete.confirmMessage'),
-    {
-      confirmButtonText: t('common.confirm'),
-      confirmButtonClass: 'danger',
-    },
-  )
-    .then(() => {
-      WorkspaceApi.deleteWorkspace(item.id as string, loading).then(async () => {
-        MsgSuccess(t('common.deleteSuccess'))
-        await getWorkspace()
-        currentWorkspace.value = item.id === currentWorkspace.value?.id ? list.value[0] : currentWorkspace.value
+async function check(id: string) {
+  try {
+    return await WorkspaceApi.deleteWorkspaceCheck(id);
+  } catch (error) {
+    console.log(error);
+  }
+}
+
+async function deleteWorkspace(item: WorkspaceItem) {
+  // 判断是否能删除
+  const canDelete = await check(item.id as string);
+  if (canDelete) {
+    MsgConfirm(
+      `${t('views.workspace.delete.confirmTitle')}${item.name} ?`,
+      t('views.workspace.delete.confirmContent'),
+      {
+        confirmButtonText: t('common.confirm'),
+        confirmButtonClass: 'danger',
+      },
+    )
+      .then(() => {
+        WorkspaceApi.deleteWorkspace(item.id as string, loading).then(async () => {
+          MsgSuccess(t('common.deleteSuccess'))
+          await getWorkspace()
+          currentWorkspace.value = item.id === currentWorkspace.value?.id ? list.value[0] : currentWorkspace.value
+        })
       })
-    })
-    .catch(() => {
-    })
+  } else {
+    MsgConfirm(
+      `${t('views.workspace.delete.confirmTitle')}${item.name} ?`,
+      t('views.workspace.delete.confirmContentNotDelete'),
+      {
+        showConfirmButton: false,
+        cancelButtonText: t('common.close'),
+      },
+    )
+  }
 }
 </script>
